+++ client/src/App.scss
@@ -192,6 +192,11 @@ Item
   padding-left: 2 * $unit;
 }
 
+.add-account__button {
+  font-size: 2 * $unit;
+  width: 100%;
+}
+
 .item-info {
   display: grid;
   grid-template-columns: 3fr 2fr 2fr;
@@ -276,6 +281,7 @@ Dev configs
 
   .admin-ledger__button {
     margin-top: 2 * $unit;
+    margin-bottom: $unit;
   }
 
   .admin-note {
@@ -373,6 +379,9 @@ Ledger
 }
 
 .return-to-payments {
+  position: absolute;
+  right: 0;
+  margin-top: 2 * $unit;
   display: flex;
   align-items: center;
   .chevron {
