+++ client/src/services/api.tsx
@@ -16,7 +16,10 @@ const api = axios.create({
 api.interceptors.response.use(
   response => response,
   error => {
-    if (error?.response?.data?.message && error.response.status === 400) {
+    if (
+      error?.response?.data?.message &&
+      (error.response.status === 400 || error.response.status === 429)
+    ) {
       throw new Error(error.response.data.message);
     }
     throw error;
